TFS log: do not store filesystem metadata in the first "extension"
This commit changes the first TFS log "extension" to only contain,
beside initial header data, a link to the next extension. This will
allow to switch from one log to another by only changing the link
in the first "extension", as will be done when log compaction is
implemented. The first "extension" now occupies just one disk
sector.
The log_new() and log_ext_new() helper functions are now called in
log_create(), and will be reused in a future commit.