TFS: implement log compaction
Any filesystem operation done by a running application causes one
or more entries to be appended to the TFS log, and these new
entries may obsolete existing log entries. As a result, after many
filesystem operations, the TFS log may contain many obsolete
entries, which wastes storage space and makes filesystem loading at
boot inefficient.
This commit implements log compaction, which eliminates obsolete
entries and recreates a "clean" log. Log compaction is triggered
when the number of obsolete and total entries in the current log
exceed certain thresholds.
In order to account for any filesystem operation done while log
compaction is in progress, log entries created during this time
are written to both the current and the new log, so that the
filesystem remains in a consistent state regardless of whether the
log compaction completes successfully.
After a successful log compaction, any tuples that are no longer
associated to existing filesystem entries are destroyed.
In order to prevent log compaction from adding to the log
non-persistent filesystem entries, any tuple corresponding to a
non-persistent entry is marked with a "no_encode" symbol, which is
checked against in the encode_tuple() function.
In order to avoid use-after-free issues if an existing log is being
flushed during log compaction, a reference count with an associated
deallocation closure is being added to the TFS log structure:
flush_log_extension() increments the refcount, and both
flush_log_extension_complete() and log_switch_complete() decrement
the refcount.

Closes #21.