Page cache: remove pagecache parameter from TFS and PC functions
Since one page cache can support multiple volumes, only a single
global page cache is needed, and passing a pagecache argument to
filesystem_create() and pagecache-related functions is unnecessary.
Removing this parameter eliminates the need to store a pagecache
pointer in static variables (except a single static variable in
pagecache.c) and the TFS filesystem structure, and allows creating
a filesystem without specifying a pagecache handle; this will make
it easier in a future commit to create filesystems associated to
mount points.