TFS header: add UUID field (TFS version bump)
This commit adds a 16-byte UUID to the TFS header of the first TFS
log extension. The new log_hdr_parse() function (which will be
reused in a future commit) has been created to parse the TFS
header. When a new filesystem is being created, the UUID is
assigned from random numbers.
Since this is a breaking change from the previous TFS header
format, the TFS version number is being bumped.
The filesystem UUID will be used to identify volumes when support
for multiple volumes (and filesystems) will be added.
The mkfs and dump tools now display in the terminal the UUID of the
filesystems being created and dumped, respectively.