mkfs/dump: add support for raw filesystem images
With this change, if a boot image path is not supplied to mkfs,
this tool does not require a kernel image or a boot filesystem, and
can create a raw (non-bootable) filesystem image (which can be
mounted by a running kernel when support for multiple filesystem
mounts is added).
In addition, if the -e command line option is specified, mkfs does
not take a manifest file as input, and creates an empty filesystem.
A filesystem image file is always padded to a multiple of the size
of TFS log extensions, otherwise if a log extension is allocated at
the end of the storage space and is not completely filled with file
metadata (as happens when an empty filesystem is created), the
extension cannot be read back when filesystem contents need to be
retrieved.

The dump tool also now supports raw filesystem images.