Storage: add support for multiple filesystem mounts
Multiple filesystems can now be mounted by specifying a "mounts"
entry in the root tuple of the manifest file. This entry is a tuple
containing one entry for each mount, with format
<uuid>:<mountpoint>, where <uuid> is a volume identifier (which
corresponds to the UUID of the filesystem contained in the volume)
and <mountpoint> is the path of the directory in which the
filesystem is to be mounted (this directory must exist in the root
filesystem). Example to mount 2 filesystems (one at /mnt and the
other at /mnt2):
mounts:(
  2b8750f4-07f2-2584-7dcc-892b3b52bd07:/mnt
  f6f1d240-6b80-2f01-3a87-186b64b7caa6:/mnt2
)
The UUID is assigned to a given filesystem when the filesystem is
created by the mkfs tool, and its value is displayed in the
terminal output by this tool (as well as by the dump tool).
In order to create a (non-bootable) filesystem to be used for a
volume, the mkfs tool must be invoked without passing a boot image
or a kernel image. To create an empty filesystem:
mkfs -e my_volume.raw
To create a non-empty filesystem:
mkfs my_volume.raw < my_volume.manifest
The manifest file in the above command must contain all the files
and folders that need to go in the filesystem.

The kernel probes any detected storage devices at boot time: if a
device contains an MBR with a partition table, then it is assumed
to be the root volume, otherwise if a filesystem is detected in the
first disk sector and its UUID corresponds to one of the mount
entries, it is mounted; if no MBR and no filesystem is found, then
the storage device is ignored.
If one or more filesystems are detected and need to be mounted, the
kernel waits for the mount operations to complete before starting
the user process.
At runtime, the filesystem corresponding to each open file is
referenced in the new "fs" field of struct file.
The virtIO SCSI driver has been modified to be able to handle
multiple SCSI targets (which are used on a Google Cloud instance
when multiple disks are attached).

Closes #472.