Unix domain sockets: implement scatter-gather I/O
This commit adds an implementation of the sg_read abd sg_write
closures to Unix domain sockets. To properly handle data that may
be accessed after it has been dequeued from a socket (as happens
when an S/G read is done), the type of queued data has been changed
from a buffer to a refcounted buffer (shareduf structure).
The bottom-half closures for reading from and writing to a socket
have been enhanced to support S/G I/O in addition to linear I/O.