syscalls: add syscall_return() helper function
This function sets the return value from a syscall and possibly
wakes up the thread for which the syscall was executed.
This allows us to get rid of boilerplate code in a few places.
file_op_*() functions have been nuked, so the only functions needed
to block and resume a thread in a syscall are
thread_maybe_sleep_uninterruptible() and syscall_return().