sigaction: add support for SA_RESTART flag
When this flag is specified in a signal action, if the signal is
raised while the recipient thread is blocked in a restartable
syscall, the syscall is restarted internally by the kernel instead
of returning -EINTR to userspace.
Restartable syscalls are those that don't have timeout values in
their argument list (or are called with an infinite timeout).
The removal of `iour->shutdown = true;` from the iour_close_bh()
function is a bug fix that was necessary even before this commit
but becomes more important now that the close() syscall can be
restarted.
The -EINVAL return value by poll(), select() and epoll_wait() was
wrong, and is being fixed by returning -EINTR (or restarting the
syscall) instead.

Closes #930.