io_getevents(): fix handling of zero-valued timeout
A zero-valued timeout (i.e. a timespec struct with both tv_sec and
tv_nsec set to 0) passed to io_getevents() means that the syscall
must just poll for any completed I/O operations and return
immediately. In the existing code, a zero-valued timeout was being
treated as no timeout at all. Fix it (and add a test case for it).