Misc fixes in preparation for enabling SMP support
Since the kernel lives at the top of the 64-bit address space, the
GDT address used at runtime does not fit in 32 bits, so a
"temporary" GDT is needed in the AP boot page for secondary CPUs to
use when switching to 64-bit mode.
Handling of the /sys/devices/system/cpu/online special file has
been updated to report the correct number of available CPUs (its
events callback has been changed so that the EPOLLIN flag is always
returned, which is the correct behavior for files that do not block
when read).