Kernel shutdown: do not disable CPUs
According to commit e353bee9cbe2dd6189d217273fc4f1568c09c811, the
shutdown interrupt was added to try to clean up crash logs. As
such, it should only be used in case of abnormal termination;
disabling a CPU when a storage sync needs to be done can lead to a
stall (because interrupts from storage devices may be routed to a
disabled CPU) that prevents the VM from being shut down.