Go runtime tests: do not fail if unknown env vars are found
https://github.com/nanovms/ops/pull/637 modified Ops so that it
automatically adds NANOS_VERSION and OPS_VERSION environment
variables to an image being created. The Go runtime test for
environment variables was not prepared to find variables other than
those specified when building the image.
This commit changes the runtime test so that it only checks for the
existence and correctness of the specified variables, and doesn't
fail if other variables are found.

Closes #1295.