Fix handling of current thread in kernel context
If SMP is enabled, a kernel context can be moved from one CPU to
another, thus the reference to the current thread should be stored
in the kernel context rather than in a per-CPU data structure such
as cpuinfo.
The get_current_thread() and set_current_thread() helper functions
are now used to retrieve and set the current thread in the kernel
context on which the current CPU is running.
In order to correctly handle page faults in kernel mode, the
current thread must be set in the kernel context before executing
I/O operations on user buffers. To do so, Unix code calls
thread_resume() (which also handles accounting of system time spent
for thread execution). As a result of this change, whenever a
blockq action is called, the current thread always coincides with
the thread associated to the action: this required a small change
in the iplementation of the futex BH handler, which otherwise would
never unblock a blocked thread unless a timeout condition occurs.
thread_pause() sets the current context to NULL (this is necessary
to avoid situations where a thread terminated in a given kernel
context is still referenced from another kernel context).
thread_log_internal() and syscall_notrace() have been modified to
avoid dereferencing the current thread (which may be NULL) and use
the thread pointer passed to thread_log_internal() instead.

Closes #1271.