TFS header: add label field (TFS version bump)
It is now possible to assign a label to a filesystem being
created. The maximum supported length for a label is 31 bytes
(plus the string terminator), and non-ASCII strings (UTF8-encoded)
are supported.
mkfs has an additional command line option (-l) that allows to
specify a label (if not specified, the default label is an emtpy
string), and the dump tool displays the label retrieved from an
existing filesystem.
The filesytem_create() function prototype has been changed to take
a label instead of a boolean "initialize" argument: a non-NULL
label indicates that a new filesystem is being created, while a
NULL label is used to read an existing filesystem.