Volume mounting: add support for volumes identified by a label
In volume mount directives included in a manifest file, it is now
possible to specify a volume by its filesystem label (alternatively
to using the filesystem UUID). This allows to easily change volumes
to be mounted by a VM instance without re-creating the image for
the instance.
Example of a mount directive using a label:
mounts:(my_label:/mnt)

The buffer_compare_with_cstring() function has been modified to be
able to handle non-ASCII UTF8-encoded strings.