PCI: add support for a driver to operate more than one device
With this change, the core PCI code allocates a struct pci_dev for
each discovered device, and keeps track of whether a driver has
been attached to it; in this way, the probe function is not called
more than once for the same device, while allowing a given driver
to control more than one device. This is needed to support multiple
storage devices instantiated as separate PCI devices.
Since now each PCI device has an allocated structure in the core
code, there is no need for drivers to keep a local struct pci_dev.