clock_gettime(): fix thread and process time accounting
Commit 29c970f955e85385882d07b1cfb1afdf1605c9f8 introduced a
regression which caused CPU time for the threads of a process to be
accounted twice in the CPU time of the process. This commit removes
the utime and stime members from struct process, which are not
necessary because the times of the individual threads need to be
summed anyway when the process CPU time is requested from
userspace.
In addition, calling set_current_thread() directly from
run_thread_frame() messes up time accounting, because when
thread_enter_user() is called it sees that the provided thread is
already the current thread. This is now fixed.

fetch_and_add_64() is now unused and is being removed; the existing
fetch_and_add() for x86 is being changed to use the same
implementation as fetch_and_add_64(), because, as noted in
https://github.com/nanovms/nanos/pull/1171#discussion_r423757729,
the compiler-generated code is a little better.