scheduler: update_timer(): fix check for maximum timer value
The existing check is failing to enforce the maximum timer value of
runloop_timer_max. As a consequence, when the system is idle,
scheduling interrupts occur as a consequence of periodic system
timers (such as lWIP timers, which have a higher interval than the
runloop_timer_max). This commit fixes the issue.