x86 APIC: remove get_cpuinfo() function
This function queries the local APIC to retrieve the current CPU,
which is considerably slower than calling current_cpu().
Retrieving the current CPU from the local APIC is only necessary
when a vCPU is started, thus ap_start() has been changed to call
directly apic_id(), and get_cpuinfo() has been removed so that it's
not called by mistake again (as was called in kernel_sleep()).