syscall_schedule(): add option to unconditionally take kernel lock
On SMP systems, when a syscall is ongoing on a core and another
core wants to execute another syscall, there are certain workloads
where it is better (performance-wise) for the second core to wait
to acquire the kernel lock instead of putting the syscall on the
runqueue and then executing the runloop, because chances are that
the core will go immediately to sleep and will be waken up again
when the syscall can be serviced or the calling thread can be
resumed (wakeup IPIs are somewhat expensive).

This commit adds a "syscall_defer" option for the manifest file: if
this option is not present, a CPU executing a syscall waits
unconditionally to acquire the kernel lock, otherwise if the kernel
lock is already taken the syscall execution is put on the runqueue.
Running the kernel without the "syscall_defer" option brings a 30%
performance improvement when benchmarking a Go web server on a
4-vCPU GCloud instance.