Scheduler: implement per-CPU thread queues
Having per-CPU thread queues instead of a single global queue
allows to better exploit local CPU caches in SMP systems, and has
been found to improve performance of a Go web server (in terms of
number of requests per second) by 25% on a 4-core GCloud instance.
In order to minimize the frequency of wakeup IPIs, idle CPUs are
only woken up if their thread queue is non-empty.
run_thread_frame() now sets the queue pointer in a thread frame, so
that after a thread is migrated from a CPU to another,
schedule_frame() will enqueue the relevant thunk to the new CPU.