SMP: add per-CPU scheduler timers
Now that each CPU has its own thread queue, in order to ensure that
a running thread does not starve other threads in the same CPU
queue we need to ensure that each CPU has its local timer armed.
Instead of using a periodic tick timer, CPUs run in tickless mode:
before running a user thread, each CPU checks its local timer, and
if the timer is not armed it is configured to fire after the
scheduling period. (Using a periodic tick timer does not appear to
bring major benefits, because (beside requiring a periodic timer
implementation in all platforms we support) the local timer would
still need to be re-configured every time a timer other than the
scheduling timer is set.)