net/direct.c: add support for client operation
Client socket connections are now supported via the
direct_connect() function. The existing code has been refactored to
reuse parts common to client and server operation.

In addition, the existing code was leaking the direct_conn struct
and the direct_conn_send closure on a connection close. This commit
fixes this leak as well, by embedding the closure in the
direct_conn struct and adding a direct_conn_dealloc() function
which is called when a connection is closed by either end.

The buffer_handler receive closure is now invoked (with a null
buffer) at every connection shutdown, regardless of which end
initiated the shutdown. This allows an application that closes a
connection to be notified when the connection is actually closed
(after which the receive closure is guaranteed to not be invoked in
the future).