klibs: add fix allocation (and deallocation) of klib virtual memory
The new walk_elf() function parses an ELF file, extracts from it
the segments to be loaded in memory, and invokes the supplied range
handler with the memory ranges of these segments.
The klib loading code was allocating only 4 KB of virtual memory
for each klib, regardless of how much memory is actually needed;
this was causing the virtual address spaces of different klibs to
overlap.
This commit fixes that by calling walk_elf() first to calculate the
total memory needed, and then allocating that amount.
In addition, it adds a missing deallocation during klib unloading.