ACPI: add detection of power button event (x86_64 only)
This commit implements detection of the ACPI power button event
(which is generated when a Google cloud instance is being stopped)
via the PIIX4 ACPI-compatible power management interface.
Upon detecting this event, the kernel shuts down, and when the
shutdown is completed the kernel puts the system in the ACPI Soft
Off state (which notifies the hypervisor that the powerdown event
has been acted upon).
Note: on AWS, even though the PIIX4 ACPI interface is emulated, an
instance stop command does not generate an ACPI power button event.
Instead, it writes the "poweroff" string in the "control/shutdown"
domU XenStore node. Thus, to implement graceful shutdown in AWS,
the Xen driver needs to support asynchronous notifications on the
XenStore event channel.

Closes #1065, #1318.