Virtqueue: make message enqueuing SMP-safe
When a message is enqueued to a virtqueue, it is placed in the
msg_queue list. Since this can happen concurrently on different
vCPUs (e.g. when doing disk I/O), the msg_queue list should be
protected against concurrent access.
This commit uses the virtqueue spinlock to make the vqmsg_commit()
function SMP-safe. virtqueue_fill() is now called with the lock
held.