Add first_boot() function
This function returns a boolean value that indicates whether this
is the first boot of the kernel on a given image.
At the first boot, the "booted" symbol is written in the filesystem
log for the root filesystem root tuple. In order for this symbol to
be correctly retrieved at a subsequent boot, the "null_value"
buffer is being changed so that it has a non-zero length
(otherwise, when reading the filesystem log, a zero-length buffer
is interpreted as an entry that removes a symbol from a given
tuple).