Radar klib: add crash reporting functionality
Crash reporting to the radar server is enabled when the manifest
file contains the RADAR_KEY environment variable.
A "crash" is defined as either a user application fatal error (e.g.
anything that causes the application to terminate with a non-zero
exit code), or a kernel fatal error. When any of these happens,
before shutting down the VM the kernel dumps on disk a trace of log
messages (printed by both the user application and the kernel); at
the next boot, the radar klib detects the log dump and sends it to
the radar server as a crash report; the value of the "id" field of
a crash report is the same as the "boot id" (sent right after each
boot), so that a crash can be unequivocally associated to a boot.

The log dump is saved in a section of the disk that is being carved
between stage2 and the boot filesystem.

Throughout the kernel, most calls to console() have been replaced
by calls to a new function, rputs(), which prints the supplied
string to both the console and the kernel log (so that in case of a
crash the console messages will be sent in the crash report).
Analogously, the buffer_print() and stdout() functions have been
changed so that their output goes to both the console and the
kernel log.