vm_exit(): add support for platform-specific VM halt function
If the platform on which the kernel is running has a defined
procedure to stop an instance, this procedure should be used not
only when the instance is stopped from the cloud platform but also
when the program finishes executing (unless the reboot_on_exit
manifest flag is set).
This commit adds a vm_halt global closure that can be set by
platform-specific code and is called by vm_exit().