TFS log: fix FS creation when log size exceeds one extension
Commit 25581ed7c75bf2915a36119e1ec893e4a8f33a94 modified
log_write() and log_write_eav() to return false (i.e. error) when
the staging buffer cannot be flushed to storage (typically, when
the storage medium is full). The implementation relies on the fact
that a flush is triggered immediately by log_set_dirty() when the
length of the staging buffer reaches a certain threshold (so that
the staging buffer will never grow beyond the size of a log
extension). This immediate triggering was only implemented in
the kernel code, and not in the non-kernel code: as a result,
filesystem creation by mkfs was failing when the log size exceeds
the size of a log extension.
This commit fixes the above issue.