netsock_listen(): do not call tcp_err() on listening socket
tcp_err() is supposed to be called only on tcp_pcb structures that
handle an actual connection, which is not the case for the
structure returned by tcp_listen_with_backlog(). In fact,
tcp_listen_with_backlog() returns a tcp_pcb_listen structure, cast
to tcp_pcb. The tcp_pcb_listen structure does not have a field for
storing an error callback, and trying to set this field results in
a memory write past the end of the structure, leading to memory
corruption.
This bug has been found to cause random runtime errors such as
"mcache ffff800000a00280: can't find cache for object
ffff800004bea800, size ffffffffffffffff; leaking".