Backed heap: add alloc_map() and dealloc_unmap() functions
alloc_map() allocates memory from both the virtual and the physical
heap, and returns to the caller both the virtual and the physical
address; dealloc_unmap() takes as arguments both the virtual and
the physical address to deallocate.
These new functions can be used as alternatives to alloc() and
dealloc() when the caller wants to use the physical address, and
allow avoiding calls to physical_from_virtual() (both when passing
a physical address to a device, and when deallocating memory),
which show up in ftrace logs as taking a significant amount of
time.
In addition, the backed heap now implements internal locking, which
can be enabled when a heap is created, so that both locking and
non-locking versions can be used.