virtIO: vqmsg_push(): take physical address as argument
This allows avoiding calls to physical_from_virtual() when the
vqmsg_push() caller already knows the physical address to be used
in a virtqueue message. The SCSI, storage and network virtIO
drivers now use the new alloc_map() and dealloc_map() functions of
the backed heap, so they can pass to vqmsg_push() the physical
addresses being allocated.