Unix process: cache lookup results for symbols in root tuple
The "trace" symbol was being looked up in the process root tuple at
every call to thread_log(); similarly, the "debugsyscalls" symbol
was being looked up at every syscall execution.
Change this so that these lookups are done only once at
initialization, in an effort to decrease the number of calls to
table_find(), which show up in ftrace logs as taking a significant
amount of time.