TCP sockets: allow reading from socket after connection closure
If data has been received from a TCP socket and is buffered in the
kernel, the application should be able to read this data even after
the remote peer closes the connection (the read() syscall should
return 0 only after all buffered data has been read).
The existing code was making read() return 0 unconditionally for a
socket whose TCP connection has been closed by the remote peer,
even if there is buffered data to be read. This commit fixes the
issue.