stage1: minimize number of INT 13H BIOS calls to load stage2
For some reason, issuing more than one INT 13H call in stage1 when
running on Xen locks up the VM (the BIOS never returns from the
second INT 13H call).
This change minimizes the number of INT 13H calls needed to load
stage2, by setting the dap.offset field to 0 and adjusting
dap.segment to align with the beginning of stage2: in this way,
stage1 can issue INT 13H calls with a sector count up to 0x80 even
if the destination buffer is not aligned to a multiple of 0x10000,
while still satisfying the constraint that the buffer fits in a
single segment.
This change allows booting Nanos on AWS t2 instances, on the
condition that the stage2 size does not grow beyond 64 kB.