Kernel clock: add NTP synchronization
This change adds a new klib that periodically retrieves the time
from an NTP server and synchronizes the kernel clock.
Clock synchronization is done gradually (by applying calibration
parameters set via clock_adjust()) so as not to create any
discontinuity in the wallclock time. The CLOCK_ID_MONOTONIC clock
is now different from CLOCK_ID_MONOTONIC_RAW, and is calculated by
applying calibration values obtained when comparing the local time
with the time retrieved via NTP. Since retrieving the monotonic
time is now slightly less efficient because of the calibration
step, various pieces of code in the kernel have been converted to
use CLOCK_ID_MONOTONIC_RAW instead.
When the local time is adjusted via clock_adjust(), the kernel
timer queue is re-ordered so that any changes in timer expiry times
resulting by clock calibration parameter changes are taken into
account. In addition, the RTC is updated to reflect the correct
time.
remove_timer() is being changed to return a timestamp referring to
the clock ID used by the timer being removed, so that the timer can
be re-registered correctly by blockq_transfer_waiters() using the
same clock ID as the original timer.