NTP: silence unaligned packed member address compiler warnings
Clang compiler emits warnings such as "taking address of packed
member 'transmit_ts' of class or structure 'ntp_packet' may result
in an unaligned pointer value". But in fact all the ntp_ts structs
packed in the ntp_packet struct are 4-byte aligned, since they are
at offsets from the beginning of the structure that are multiple of
4 bytes, thus their u32 members are always at aligned addresses.
This change casts the struct ntp_ts pointers to void* to silence
the warnings.