NTP: add dynamic adjustment and configurability of poll interval
The NTP poll interval is now adjusted dynamically between a minimum
and a maximum value, based on the measured jitter of offset values
between the local time and the NTP time, using an algorithm similar
to that of the NTP reference implementation, whose purpose is to
maximize clock precision while minimizing network traffic overhead.
The minimum and maximum poll interval values are expressed as bit
order of the number of seconds, and are set by default to 6 (i.e.
64 seconds) and 10 (i.e. 1024 seconds) respectively. These values
are configurable via the "ntp_poll_min" and "ntp_poll_max" manifest
options, respectively, whose allowed values range from 4 (i.e. 16
seconds to 17 (i.e. 36.4 hours).