x86_43: cpu_init(): initialize CR4 and CR0 control registers
When the kernel ELF file is loaded directly by the hypervisor in
long mode, the boot process does not go through stage1 and stage2.
Thus, appropriate settings for the CR4 and CR0 registers of the
bootstrap processor must be enforced in the kernel code.
This change moves initialization of CR4 and CR0 from the
PREPARE_LONG_MODE and ENTER_LONG_MODE assembly macros to the
cpu_init() function. Now the assembly macros only do what is
strictly necessary to enter long mode.
This fixes UD2 exceptions being triggered when running under AWS
Firecracker.