Go tests: remove references to Mkfs and set base volume size
With https://github.com/nanovms/ops/pull/887, Ops no longer uses
the mkfs tool to create image files, and for this reason the
lepton.Config struct no longer contains the Mkfs field.
In addition, when creating an image whose program needs to write to
the filesystem, a base volume size has to be configured, otherwise
there is no free space in the filesystem to create new files.