netsyscall.c refactoring
A few utility functions are now implememnted in net.c and declared
in a header file, so that they can be used from outside
netsyscall.c.
In addition, the getsockname function pointer is being added to
struct sock, so that different socket types can implement this
functionality.
This refactoring is done in preparation for adding support for
netlink sockets.