ID heap set_area(): update count of allocated resources
If set_area() is called with `validate` set to `true` and the
validation succeeds, then it should update the `allocated` struct
member, to keep a consistent count of allocated resources.
This is necessary to correctly measure disk space utilization,
because filesystem entries read from the TFS log are accounted for
in the storage heap by calling id_heap_set_area().