Klibs: radar: add sending of disk space utilization metrics
Disk space usage data are sent to the same server endpoint
(/api/v1/machine-stats) as used for memory usage data, and with the
same periodicity (i.e. 5 minutes). The JSON string sent to the
endpoint now contains an additional "diskUsage" attribute, whose
value is an array of JSON objects (one for each disk mounted by the
instance); each array element contains 3 attributes:
- "volume": a string that identifies the volume; for the root
volume, the string value is "root"; for any additional volumes, the
string value is the volume label if present, or the volume UUID if
the label is not present
- "used": the number of bytes used by the filesystem (file contents
and meta-data) in the volume
- "total": the total number of bytes in the storage space of the
volume, i.e. the upper limit for the "used" attribute value

Example of a JSON string containing both memory and disk usage
data:
```
"{"bootID":1,"memUsed":[112705536,112705536,112705536,112705536,
112705536],"diskUsage":[{"volume":"root","used":10438656,"total":
12582912},{"volume":"mnt","used":107836416,"total":232783872}]}"
```
In the above example, the root volume usage is ~10 MB out of a
total of ~12 MB, and there is an additional volume, labeled "mnt",
whose filesystem occupies ~107 MB out of ~230 MB.

Example of a substring representing a volume without a label,
which is thus identified by its UUID:
```
{"volume":"2375a5a1-2d36-15cf-ea6b-2fa01df2ccde","used":524800,
"total":5248000}
```