Virtio-net: add check for VIRTIO_F_ANY_LAYOUT feature
The VIRTIO_F_ANY_LAYOUT feature, when using the legacy virtio
interface, indicates that no assumptions are made about message
framing.
For network devices, the virtio spec says that "drivers which have
not negotiated VIRTIO_F_ANY_LAYOUT MUST use a single descriptor for
the struct virtio_net_hdr on both transmit and receive, with the
network data in the following descriptors". This part of the spec
was implemented for the trasmit queue, but not for the receive
queue.
With this change, the virtio-net driver offers the
VIRTIO_F_ANY_LAYOUT feature at initialization, and checks whether
the device accepts it; if the feature is not accepted, the first
descriptor of a packet in the receive queue is dedicated to the
struct virtio_net_hdr, and a second descriptor is used for the
network data.
This change is necessary to support the virtio-net device as
implemented in VirtualBox.