ATA drive probe: wait for BUSY flag to be cleared
When running under VirtualBox, the ATA drive may be in a busy state
when it is being probed; if the IDENTIFY command is issued when the
drive is in this state, the drive does not process the command
properly, and as a result the probing procedure fails.
This change adds a call to ata_wait() in the probe code, so that
the drive is guaranteed to not be busy when the IDENTIFY command is
issued.