Explicitly compile virt kernel with -fno-pic
Without -fno-pic, newer versions of gcc (9+) were generating code that was
loading constants through the global offset table. As memory locations stored
in this table are absolute and based on the linked (not loaded) address, early
initialization code was generating faults while attempting to access global
variables at the linked addresses before the MMU was enabled. PIC is not
necessary for kernel code, and explicitly specifying -fno-pic eliminates
these GOT lookups in favor of PC-relative accesses.