x86_64: implement TSC calibration using PIT
This allows the kernel to run on hypervisors that don't expose a
known paravirtualization interface and don't implement the HPET.
The HPET is now used only as a last resort, if the calibrated TSC
cannot be used.
The "failed to initialize HPET; invalid femtoperiod" error message,
which would appear when probing a non-existing HPET, is being
removed because it may no longer be a fatal error.