buffered serial console output
Previously, all serial console output was written directly to the device,
polling if the transmit FIFO was full. This change adds a ring buffer to store
console output, utilizing interrupts to write characters to the device when
space is available in the transmit FIFO. This option may be disabled by
specifying "polled_serial" in the root tuple.

This change should reduce the overhead of bursty writes to the console, as
writes will no longer spin waiting space to clear in the serial transmit
FIFO (until the driver's ring buffer fills up).