load_elf: move non-page-aligned loadable bits to bss mapping
When mapping PT_LOAD segments with a bss that does not start on a page
boundary, copy preceding loadable data, up to the page boundary, into the
allocated mapping for bss. Prior to this change, the page containing such
trailing loadable bits would be mapped directly from the source buffer, and
the start of bss to the end of this page would be zeroed. Placing bss within
this page could lead to corruption of another section (e.g. .symtab) if it was
placed just after the end of the loadable data within this last page.

In lieu of the obviated vpzero macro, change elf_map_handlers to return a
virtual address where the created mapping can be accessed. This is typically
just the supplied vaddr, but in stage2 this is the identity address - which is
allocated by the handler for the bss mapping.