revert id heap default behavior to first-fit
The id heap now defaults to first-fit allocation. Previously, the id heap
would default to starting the allocation search from the location of a
previous allocation or deallocation. This behavior was leading to excessive
fragmentation of resources, as could be easily demonstrated by making
alternating requests of 2M and 4K-sized pages from the physical heap. After a
4K allocation, a subsequent 2M allocation would start at the next 2M
boundary (to satisfy the id heap alignment property), and the preceding unused
space would not be considered for allocation unless the allocation search
wrapped within the id range or the next_bit was explicitly reset using
id_heap_set_next.

Next-fit behavior may still be induced by the calling code with use of
id_heap_set_next.