x86_64: interrupt entry: fix check for kernel mode
The value of the CS register when in 64-bit kernel mode is not
necessarily 8: the only bits that have a pre-defined value are the
2 least significant bits, which contain the Current Privilege Level
(CPL).
This commit changes the check for kernel mode in the check_swapgs
assembly macro in order to check the 2 least significant bits
instead of the entire value of the CS register. This allows booting
the kernel via UEFI.