x86: init_vsyscall(): fix mapping flags for vsyscall memory
After mapping the vsyscall memory, the kernel wrties to this area,
thus the initial mapping must have read/write flags; only after the
writes are done, the mapping flags can be changed to read-only.
This commit fixes a page fault that occurs if the kernel is booted
with the WP (Write Protect) bit of the CR0 register set to 1.