x86: add UEFI bootloader
This allows Nanos to boot on UEFI platforms, such as Hyper-V
Generation-2 instances (#1416).
The UEFI bootloader file is called bootx64.efi and is located in
the ESP (EFI System Partition), in a FAT32 filesystem in the
EFI/Boot/ folder. In order to create this partition in the Nanos
image file, the -u option needs to be supplied to the mkfs tool.
The mkfs and dump tools, as well as the stage2 and kernel code,
have been modified in order to properly handle images with and
without the ESP.

For the Mac build, the objcopy tool (from the binutils package) is
being used instead of x86_64-elf-objcopy, because the latter does
not support the efi-app-x86_64 BFD target that is used to create
the UEFI bootloader file.