Htper-V Storvsc: ignore VSTOR_OPERATION_ENUMERATE_BUS messages
When running in a Generation 2 VM, after sending the "report ready"
packet to the Azure platform, the guest receives from the
hypervisor a VSTOR_OPERATION_ENUMERATE_BUS message in the Storvsc
VMBus channel. Even though the driver doesn't support
re-enumeration of the VMBus, it shouldn't halt the VM on reception
of such message.